// src/app/page/master/map-program-to-jurusan/map-program-to-jurusan.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapProgramToJurusan {
	kdProduk?: string;
	kdEvent?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}